+++ native/c/test/zutils.cpp
@@ -103,6 +103,15 @@ std::string get_random_string(const int length, const bool allNumbers)
   return ret;
 }
 
+std::string get_basename(const std::string &fspath)
+{
+  if (fspath.empty()) return "";
+
+  auto lastSlash = fspath.find_last_of("/\\");
+
+  return (lastSlash == std::string::npos) ? std::string(fspath) : fspath.substr(lastSlash + 1);
+}
+
 std::string get_random_uss(const std::string &base_dir)
 {
   std::string ret{base_dir};