+++ native/c/zusf.hpp
@@ -22,6 +22,18 @@
 #include <sys/types.h>
 #include "zusftype.h"
 
+struct CopyOptions
+{
+  bool recursive;
+  bool follow_symlinks;
+  bool preserve_attributes;
+  bool force;
+  explicit CopyOptions(bool recursive = false, bool follow_symlinks = false, bool preserve_attributes = false, bool force = false)
+      : recursive(recursive), follow_symlinks(follow_symlinks), preserve_attributes(preserve_attributes), force(force)
+  {
+  }
+};
+
 struct ListOptions
 {
   bool all_files;
@@ -33,6 +45,7 @@ struct ListOptions
   }
 };
 
+int zusf_copy_file_or_dir(ZUSF *zusf, const std::string &source_fs, const std::string &dest_fs, const CopyOptions& options);
 int zusf_create_uss_file_or_dir(ZUSF *zusf, const std::string &file, mode_t mode, bool createDir);
 std::string zusf_format_file_entry(ZUSF *zusf, const struct stat &file_stats, const std::string &file_path, const std::string &display_name, ListOptions options, bool use_csv_format);
 int zusf_list_uss_file_path(ZUSF *zusf, const std::string &file, std::string &response, ListOptions options = ListOptions{}, bool use_csv_format = false);